refactor: move ms implementations into a subpackage
