+++ kotlin-utils/src/test/kotlin/xyz/lbres/kotlinutils/set/multiset/impl/immutable/runImmutableUnaryTests.kt
@@ -1,4 +1,4 @@
-package xyz.lbres.kotlinutils.set.multiset.mutable
+package xyz.lbres.kotlinutils.set.multiset.impl.immutable
 
 import xyz.lbres.kotlinutils.list.IntList
 import xyz.lbres.kotlinutils.set.multiset.* // ktlint-disable no-wildcard-imports no-unused-imports
@@ -6,54 +6,33 @@ import kotlin.test.assertEquals
 import kotlin.test.assertFalse
 import kotlin.test.assertTrue
 
-fun runMutableIsEmptyTests() {
+fun runImmutableIsEmptyTests() {
     // empty
-    var intSet: MutableMultiSet<Int> = mutableMultiSetOf()
+    var intSet: MultiSet<Int> = multiSetOf()
     assertTrue(intSet.isEmpty())
 
-    var stringSet: MutableMultiSet<String> = mutableMultiSetOf()
+    var stringSet: MultiSet<String> = multiSetOf()
     assertTrue(stringSet.isEmpty())
 
     // not empty
-    intSet = mutableMultiSetOf(0)
+    intSet = multiSetOf(0)
     assertFalse(intSet.isEmpty())
 
-    intSet = mutableMultiSetOf(1000, -1000, 4, 2, 4)
+    intSet = multiSetOf(1000, -1000, 4, 2, 4)
     assertFalse(intSet.isEmpty())
 
-    intSet = mutableMultiSetOf(3, 3, 3)
+    intSet = multiSetOf(3, 3, 3)
     assertFalse(intSet.isEmpty())
 
-    stringSet = mutableMultiSetOf("123", "abc")
+    stringSet = multiSetOf("123", "abc")
     assertFalse(stringSet.isEmpty())
 
-    stringSet = mutableMultiSetOf("abcdefg", "abcdefg")
+    stringSet = multiSetOf("abcdefg", "abcdefg")
     assertFalse(stringSet.isEmpty())
-
-    // remove elements
-    intSet = mutableMultiSetOf(1)
-    intSet.remove(1)
-    assertTrue(intSet.isEmpty())
-
-    intSet = mutableMultiSetOf(1, 1)
-    intSet.remove(1)
-    assertFalse(intSet.isEmpty())
-    intSet.remove(1)
-    assertTrue(intSet.isEmpty())
-
-    intSet = mutableMultiSetOf(2, 3)
-    intSet.remove(3)
-    assertFalse(intSet.isEmpty())
-    intSet.remove(2)
-    assertTrue(intSet.isEmpty())
-
-    intSet = mutableMultiSetOf(2, 3)
-    intSet.clear()
-    assertTrue(intSet.isEmpty())
 }
 
-fun runMutableGetCountOfTests() {
-    var set: MutableMultiSet<Int> = mutableMultiSetOf()
+fun runImmutableGetCountOfTests() {
+    var set: MultiSet<Int> = multiSetOf()
     var expected = 0
 
     var value = 0
@@ -62,7 +41,7 @@ fun runMutableGetCountOfTests() {
     value = 100
     assertEquals(expected, set.getCountOf(value))
 
-    set = mutableMultiSetOf(2)
+    set = multiSetOf(2)
 
     value = 2
     expected = 1
@@ -72,7 +51,7 @@ fun runMutableGetCountOfTests() {
     expected = 0
     assertEquals(expected, set.getCountOf(value))
 
-    set = mutableMultiSetOf(1, 1, 2, 1, -4, 5, 2)
+    set = multiSetOf(1, 1, 2, 1, -4, 5, 2)
 
     value = 1
     expected = 3
@@ -92,7 +71,7 @@ fun runMutableGetCountOfTests() {
 
     val mutableList1 = mutableListOf(1, 2, 3)
     val mutableList2 = mutableListOf(1, 2, 3)
-    val listSet: MutableMultiSet<IntList> = mutableMultiSetOf(mutableList1, mutableList2)
+    val listSet: MultiSet<IntList> = multiSetOf(mutableList1, mutableList2)
 
     var listValue = listOf(1, 2, 3)
     expected = 2
